# loan-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application API",
        "contact": {
            "email": "lehongduc87@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Access token",
                                                    "type": "string",
                                                    "example": "2|MBsbYZCr6icrFpy3oJYZPqDvnyRCOHuuZtZcZw0E",
                                                    "nullable": false
                                                },
                                                "token_type": {
                                                    "description": "Token type",
                                                    "type": "string",
                                                    "example": "Bearer",
                                                    "nullable": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "operationId": "ae2f8c1ae1f9c5aee99c01f5d13da30b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/loans": {
            "get": {
                "tags": [
                    "Loan"
                ],
                "operationId": "cc1f90499cec1cbd6918f5b05a9a3248",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LoanSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loan"
                ],
                "operationId": "0d43f12ab566cb3ad27ea972d4fc5be2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLoanRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LoanSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/loans/{id}": {
            "get": {
                "tags": [
                    "Loan"
                ],
                "operationId": "70df0957fb515923adb99d2f5c01f6f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LoanSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/loans/{id}/approve": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "operationId": "b07b02f401988356abbcd40b6610b423",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Loan ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LoanSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payments": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "operationId": "46ac3ded6ff7051619669d480b36714a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PayRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LoanSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateLoanRequestSchema": {
                "properties": {
                    "amount": {
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "10000",
                        "nullable": false
                    },
                    "terms": {
                        "description": "Terms",
                        "type": "integer",
                        "example": "3",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "LoginRequestSchema": {
                "properties": {
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "lehongduc87@gmail.com",
                        "nullable": false
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "example": "password",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "PayRequestSchema": {
                "properties": {
                    "installment_id": {
                        "description": "Installment ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "amount": {
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "3333.33",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "InstallmentSchema": {
                "properties": {
                    "id": {
                        "description": "Installment ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "loan_id": {
                        "description": "Loan ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "installment_no": {
                        "description": "Installment number",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "amount": {
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "3333.33",
                        "nullable": false
                    },
                    "due_date": {
                        "description": "Due date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-31T08:18:55.000000Z",
                        "nullable": false
                    },
                    "paid_date": {
                        "description": "Due date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-31T08:18:55.000000Z",
                        "nullable": true
                    },
                    "paid_amount": {
                        "description": "Paid amount",
                        "type": "string",
                        "format": "date-time",
                        "example": "3333.33",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status",
                        "type": "string",
                        "enum": [
                            "1 - Pending",
                            "2 - Paid"
                        ],
                        "example": "1",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-31T08:18:55.000000Z",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-31T08:18:55.000000Z",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "LoanSchema": {
                "properties": {
                    "id": {
                        "description": "Loan ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "amount": {
                        "description": "Amount",
                        "type": "number",
                        "format": "float",
                        "example": "10000",
                        "nullable": false
                    },
                    "terms": {
                        "description": "Terms",
                        "type": "integer",
                        "example": "3",
                        "nullable": false
                    },
                    "status": {
                        "description": "Status",
                        "type": "string",
                        "enum": [
                            "1 - Pending",
                            "2 - Approved",
                            "3 - Canceled",
                            "4 - Paid"
                        ],
                        "example": "1",
                        "nullable": false
                    },
                    "created_at": {
                        "description": "Created at",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "Updated at",
                        "type": "string",
                        "format": "date-time",
                        "nullable": false
                    },
                    "installments": {
                        "description": "List of installments",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InstallmentSchema"
                        }
                    }
                },
                "type": "object"
            },
            "UserSchema": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": "1",
                        "nullable": false
                    },
                    "name": {
                        "type": "string",
                        "example": "Duc Le",
                        "nullable": false
                    },
                    "email": {
                        "type": "string",
                        "example": "lehongduc87@gmail.com",
                        "nullable": false
                    },
                    "type": {
                        "type": "integer",
                        "enum": [
                            "1 - admin",
                            "2 - user"
                        ],
                        "example": "2",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}